PCHR-3084: Move backstopjs tasks in a separate module
